[Feature #158628185],[Feature #158628186] Add tests for creating new ride offer.
